# [TraceQL] SyncIterator seekWithinPage for improved performance