--- a/pkg/parquetquery/iters.go
+++ b/pkg/parquetquery/iters.go
@@ -586,37 +586,7 @@ func (c *SyncIterator) SeekTo(to RowNumber, definitionLevel int) (*IteratorResul
 		return nil, nil
 	}
 
-	// if c.columnName == "rs.list.element.Resource.Attrs.list.element.Key" && to[0] == 1 {
-	// 	fmt.Println("?") // this is when bad things happen
-	// } else if c.columnName == "rs.list.element.Resource.Attrs.list.element.Key" {
-	// 	fmt.Println("!")
-	// }
-
-	// only skip if to is ahead of curr
-	if CompareRowNumbers(definitionLevel, to, c.curr) > 0 {
-		// skips are calculated off the start of the page
-		rowSkip := to[0] - c.currPageMin[0]
-		if rowSkip > 1 {
-			// reslice the page to jump directly to the desired row number
-			pg := c.currPage.Slice(rowSkip-1, c.currPage.NumRows())
-			pq.Release(c.currPage) //- can i release here? i think the internal buffers are  are preserved, so no
-
-			// set curr to what it will be now that we've resliced
-			c.curr = to
-			// remove all detail below the row number
-			for i := 1; i < 5; i++ {
-				c.curr[i] = -1
-			}
-			c.curr = c.curr.Preceding()
-
-			// reset buffers and other vars
-			c.currPage = pg
-			c.currPageMin = c.curr
-			c.currValues = pg.Values()
-			syncIteratorPoolPut(c.currBuf)
-			c.currBuf = nil
-		}
-	}
+	c.seekRows(to, definitionLevel)
 
 	// The row group and page have been selected to where this value is possibly
 	// located. Now scan through the page and look for it.
@@ -743,6 +713,34 @@ func (c *SyncIterator) seekPages(seekTo RowNumber, definitionLevel int) (done bo
 	}
 
 	return false, nil
+}
+
+func (c *SyncIterator) seekRows(to RowNumber, definitionLevel int) {
+	// only skip if to is ahead of curr
+	if CompareRowNumbers(definitionLevel, to, c.curr) > 0 {
+		// skips are calculated off the start of the page
+		rowSkip := to[0] - c.currPageMin[0]
+		if rowSkip > 5 { // made up threshold worth skipping
+			// reslice the page to jump directly to the desired row number
+			pg := c.currPage.Slice(rowSkip-1, c.currPage.NumRows())
+			pq.Release(c.currPage) //- can i release here? i think the internal buffers are  are preserved, so no
+
+			// set curr to what it will be now that we've resliced
+			c.curr = to
+			// remove all detail below the row number
+			for i := 1; i < 5; i++ {
+				c.curr[i] = -1
+			}
+			c.curr = c.curr.Preceding()
+
+			// reset buffers and other vars
+			c.currPage = pg
+			c.currPageMin = c.curr
+			c.currValues = pg.Values()
+			syncIteratorPoolPut(c.currBuf)
+			c.currBuf = nil
+		}
+	}
 }
 
 // next is the core functionality of this iterator and returns the next matching result. This