@stoewer 2023-11-30
I reproduced the benchmark, here are my results

Benchmark cmd
```
go test -run='^#' -bench=BackendBlockTraceQL -count=10 -benchtime=2s -benchmem ./tempodb/encoding/vparquet3
```

Results
```
goos: linux
goarch: amd64
pkg: github.com/grafana/tempo/tempodb/encoding/vparquet3
cpu: AMD Ryzen 7 PRO 6850U with Radeon Graphics     
                                                 │ bench-magic-before-01.txt │      bench-magic-after-01.txt       │
                                                 │          sec/op           │   sec/op     vs base                │
BackendBlockTraceQL/spanAttValNoMatch-16                         12.06m ± 1%   11.64m ± 1%   -3.48% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                   11.47m ± 1%   11.06m ± 1%   -3.54% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-16                     11.28m ± 0%   10.94m ± 0%   -3.01% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                 31.91m ± 1%   27.20m ± 1%  -14.75% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-16                           238.0m ± 3%   231.1m ± 2%   -2.90% (p=0.003 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                     11.27m ± 1%   11.17m ± 1%   -0.87% (p=0.005 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-16                      251.6m ± 3%   247.2m ± 3%        ~ (p=0.089 n=10)
BackendBlockTraceQL/count-16                                     398.9m ± 1%   420.5m ± 1%   +5.43% (p=0.000 n=10)
BackendBlockTraceQL/struct-16                                    706.1m ± 3%   705.0m ± 2%        ~ (p=0.853 n=10)
BackendBlockTraceQL/||-16                                       109.40m ± 1%   73.67m ± 1%  -32.66% (p=0.000 n=10)
BackendBlockTraceQL/mixed-16                                     56.67m ± 2%   39.94m ± 1%  -29.53% (p=0.000 n=10)
geomean                                                          62.68m        57.17m        -8.79%

                                                 │ bench-magic-before-01.txt │       bench-magic-after-01.txt       │
                                                 │            B/s            │     B/s       vs base                │
BackendBlockTraceQL/spanAttValNoMatch-16                        206.3Mi ± 1%   213.8Mi ± 1%   +3.60% (p=0.000 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                  209.5Mi ± 1%   217.2Mi ± 1%   +3.68% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-16                    125.5Mi ± 0%   129.4Mi ± 0%   +3.10% (p=0.000 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                460.3Mi ± 1%   539.9Mi ± 1%  +17.30% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-16                          11.58Mi ± 3%   11.92Mi ± 2%   +2.97% (p=0.003 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                    126.8Mi ± 1%   127.9Mi ± 1%   +0.88% (p=0.005 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-16                     42.52Mi ± 3%   43.27Mi ± 3%        ~ (p=0.093 n=10)
BackendBlockTraceQL/count-16                                    38.97Mi ± 1%   36.96Mi ± 1%   -5.15% (p=0.000 n=10)
BackendBlockTraceQL/struct-16                                   6.199Mi ± 3%   6.208Mi ± 2%        ~ (p=0.839 n=10)
BackendBlockTraceQL/||-16                                       145.3Mi ± 1%   215.8Mi ± 1%  +48.51% (p=0.000 n=10)
BackendBlockTraceQL/mixed-16                                    257.6Mi ± 2%   365.6Mi ± 1%  +41.90% (p=0.000 n=10)
geomean                                                         83.69Mi        91.75Mi        +9.63%

                                                 │ bench-magic-before-01.txt │      bench-magic-after-01.txt       │
                                                 │         MB_io/op          │  MB_io/op   vs base                 │
BackendBlockTraceQL/spanAttValNoMatch-16                          2.609 ± 0%   2.609 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                    2.519 ± 0%   2.519 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttValNoMatch-16                      1.485 ± 0%   1.485 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                  15.40 ± 0%   15.40 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValNoMatch-16                            2.889 ± 0%   2.889 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchAnd-16                      1.499 ± 0%   1.499 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixedValMixedMatchOr-16                       11.22 ± 0%   11.22 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/count-16                                      16.30 ± 0%   16.30 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/struct-16                                     4.589 ± 0%   4.589 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/||-16                                         16.67 ± 0%   16.67 ± 0%       ~ (p=1.000 n=10) ¹
BackendBlockTraceQL/mixed-16                                      15.31 ± 0%   15.31 ± 0%       ~ (p=1.000 n=10) ¹
geomean                                                           5.500        5.500       +0.00%
¹ all samples are equal

                                                 │ bench-magic-before-01.txt │       bench-magic-after-01.txt       │
                                                 │           B/op            │     B/op       vs base               │
BackendBlockTraceQL/spanAttValNoMatch-16                        7.995Mi ± 1%   7.982Mi ±  1%       ~ (p=0.796 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                  7.040Mi ± 1%   7.052Mi ±  1%       ~ (p=0.853 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-16                    6.967Mi ± 1%   6.989Mi ±  1%       ~ (p=0.315 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                8.535Mi ± 1%   8.701Mi ±  1%  +1.94% (p=0.002 n=10)
BackendBlockTraceQL/mixedValNoMatch-16                          8.490Mi ± 6%   8.487Mi ±  5%       ~ (p=0.971 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                    6.972Mi ± 1%   6.967Mi ±  1%       ~ (p=0.739 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-16                     7.517Mi ± 4%   7.865Mi ± 10%       ~ (p=0.165 n=10)
BackendBlockTraceQL/count-16                                    262.5Mi ± 1%   261.7Mi ±  1%       ~ (p=0.971 n=10)
BackendBlockTraceQL/struct-16                                   28.16Mi ± 8%   26.97Mi ±  5%       ~ (p=0.218 n=10)
BackendBlockTraceQL/||-16                                       15.97Mi ± 2%   16.74Mi ±  3%  +4.83% (p=0.000 n=10)
BackendBlockTraceQL/mixed-16                                    7.407Mi ± 3%   7.481Mi ±  6%       ~ (p=0.436 n=10)
geomean                                                         12.63Mi        12.72Mi        +0.71%

                                                 │ bench-magic-before-01.txt │      bench-magic-after-01.txt      │
                                                 │         allocs/op         │  allocs/op   vs base               │
BackendBlockTraceQL/spanAttValNoMatch-16                         109.5k ± 0%   109.5k ± 0%       ~ (p=0.677 n=10)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-16                   109.4k ± 0%   109.4k ± 0%       ~ (p=0.524 n=10)
BackendBlockTraceQL/resourceAttValNoMatch-16                     109.4k ± 0%   109.4k ± 0%       ~ (p=0.261 n=10)
BackendBlockTraceQL/resourceAttIntrinsicMatch-16                 116.2k ± 0%   116.4k ± 0%  +0.16% (p=0.000 n=10)
BackendBlockTraceQL/mixedValNoMatch-16                           111.8k ± 0%   111.8k ± 0%       ~ (p=0.667 n=10)
BackendBlockTraceQL/mixedValMixedMatchAnd-16                     109.5k ± 0%   109.5k ± 0%       ~ (p=0.836 n=10)
BackendBlockTraceQL/mixedValMixedMatchOr-16                      110.8k ± 0%   110.8k ± 0%  +0.02% (p=0.000 n=10)
BackendBlockTraceQL/count-16                                     1.513M ± 0%   1.533M ± 0%  +1.35% (p=0.000 n=10)
BackendBlockTraceQL/struct-16                                    331.5k ± 0%   331.4k ± 1%       ~ (p=0.529 n=10)
BackendBlockTraceQL/||-16                                        181.3k ± 0%   198.0k ± 0%  +9.21% (p=0.000 n=10)
BackendBlockTraceQL/mixed-16                                     111.2k ± 0%   112.2k ± 0%  +0.93% (p=0.000 n=10)
geomean                                                          162.5k        164.2k       +1.03%

```