@joe-elliott 2023-12-08
After finding a regression on other queries, I've improved the change to count the exact number of nexts that the seek will replace and only seeking if it passes a threshold that seems to agree with all available benchmarks. This changes the algorithm enough that I think a second review would be worthwhile (cc @mdisibio, @stoewer).

New cpu benches:
```
name                                             old time/op    new time/op     delta
BackendBlockTraceQL/spanAttValNoMatch-8            8.97ms ± 1%     8.98ms ± 0%      ~     (p=0.222 n=5+5)
BackendBlockTraceQL/spanAttIntrinsicNoMatch-8      8.51ms ± 0%     8.51ms ± 0%      ~     (p=0.841 n=5+5)
BackendBlockTraceQL/resourceAttValNoMatch-8        8.33ms ± 1%     8.36ms ± 1%      ~     (p=0.421 n=5+5)
BackendBlockTraceQL/resourceAttIntrinsicMatch-8    28.8ms ± 1%     25.4ms ± 1%   -11.64%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixedValNoMatch-8               261ms ± 3%      258ms ± 0%      ~     (p=0.421 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchAnd-8        8.36ms ± 1%     8.44ms ± 1%      ~     (p=0.056 n=5+5)
BackendBlockTraceQL/mixedValMixedMatchOr-8          275ms ± 2%      267ms ± 0%    -2.92%  (p=0.008 n=5+5)
BackendBlockTraceQL/count-8                         501ms ± 7%      514ms ±11%      ~     (p=0.841 n=5+5)
BackendBlockTraceQL/struct-8                        788ms ± 5%      776ms ± 1%      ~     (p=0.151 n=5+5)
BackendBlockTraceQL/||-8                            182ms ± 1%      109ms ± 0%   -40.44%  (p=0.008 n=5+5)
BackendBlockTraceQL/mixed-8                        40.9ms ± 1%     27.7ms ± 1%   -32.21%  (p=0.008 n=5+5)
```